feat: Inventory 동작 끝, Shop Item 선택 UX 개선, ItemIcon 적용

+++ Source/LeagueofLegends/private/UI/ViewModel/InventoryViewModel.cpp
@@ -52,5 +52,5 @@ void UInventoryViewModel::HandleInventorySlotChanged(int32 SlotIndex, UItemInsta
 
 void UInventoryViewModel::HandleGoldChanged(float NewGold)
 {
-	OnGoldChanged.Broadcast(NewGold);
+	OnGoldChanged.Broadcast(static_cast<int32>(NewGold));
 }
